Deleted unecessary SingleReponseEntity and repository. Added new tests to ContactRepositoryIT class
